inference/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.9.0
redis==5.1.0
httpx==0.27.2
Pillow==10.4.0
ultralytics>=8.3.9
torch==2.4.1
torchvision==0.19.1
numpy==1.26.4
